Added files via upload
This code takes time and date off the RTC, and outputs to an LCD and serial port. 
No provision is made for setting time, other than a default setting available.  
Verified working on PIC16F88. This code is a stripped down version of larger project, after changing from DS1302 RTC to DS3234. 
This uses a software SPI, after I had problems with the hardware SPI.